Split ITree into IDirectedGraph and IExpressionTree
